django/requirements.txt
Django
python-dotenv
mysqlclient
django-allauth
celery
redis
django-celery-beat
django-celery-results
djangorestframework
drf-spectacular
django-storages
boto3
django-cors-headers
django-hosts
django-allauth
django-autoslug
Pillow
pytz
requests
django-filter
fido2
gunicorn
django-allauth[mfa]